remove warning

+++ src/search/hdgbfs.cc
@@ -126,7 +126,7 @@ int HDGBFS::Search() {
   auto state = InitialEvaluate();
 
   if (runup() && rank() == initial_rank()) {
-    while (n_open_nodes() < world_size() && !NoNode()) {
+    while (static_cast<int>(n_open_nodes()) < world_size() && !NoNode()) {
       int node = Pop();
       int goal = IndependentExpand(node, state);
