+++ src/search/pddsgbfs.cc
@@ -10,7 +10,7 @@ int PDDSGBFS::Search() {
   auto state = InitialEvaluate(true);
 
   if (runup() && rank() == initial_rank()) {
-    while (n_open_nodes() < world_size() && !NoNode()) {
+    while (static_cast<int>(n_open_nodes()) < world_size() && !NoNode()) {
       int node = Pop();
       int goal = IndependentExpand(node, state, true);
