+++ src/search_graph/search_graph_with_timestamp.h
@@ -63,7 +63,7 @@ class SearchGraphWithTimestamp : public T {
   }
 
   void SetH(int i, int h) override {
-    if (hs_.size() <= i)
+    if (static_cast<int>(hs_.size()) <= i)
       hs_.resize(i + 1);
 
     hs_[i] = h;
